requirements.txt
# Core Django stack
Django>=4.2,<5.0
djangorestframework>=3.14.0
drf-yasg>=1.21.5

# Machine learning & data processing
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
joblib>=1.3.0

# Telegram bot integration
python-telegram-bot>=20.6,<21.0  # v20+ is asyncio-based

# Database
psycopg2-binary>=2.9.5  # Use only in development; use `psycopg2` in production

# Deployment
gunicorn>=21.2.0

# Optional: Load environment variables securely
python-dotenv>=1.0.0

# Optional: Use whitenoise to serve static files in production
whitenoise>=6.5.0